Hello Again
The thing I'm working on is the delete function
I have tried copying existing todos to a temp.txt file
and then deleting the todo.txt file and then again creating a new todo.txt file and excluding that one line.
I have also tried to rename the temp.txt file to todo.txt file but for unknown reason it is not working!
So currently I'm figuring how to do it
r